docs/using-testharness.html

<!DOCTYPE html>  <html> <head>   <title>using-testharness.js</title>   <meta http-equiv="content-type" content="text/html; charset=UTF-8">   <link rel="stylesheet" media="all" href="docco.css" /> </head> <body>   <div id="container">     <div id="background"></div>          <table cellpadding="0" cellspacing="0">       <thead>         <tr>           <th class="docs">             <h1>               using-testharness.js             </h1>           </th>           <th class="code">           </th>         </tr>       </thead>       <tbody>                               <tr id="section-1">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-1">&#182;</a>               </div>               <p><!--</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="cm">/*global test assert_true assert_false assert_equals assert_not_equals assert_in_array</span>
<span class="cm">         assert_array_equals assert_approx_equals assert_regexp_match assert_own_property</span>
<span class="cm">         assert_inherits assert_idl_attribute assert_readonly assert_throws assert_unreached</span>
<span class="cm">         assert_object_equals*/</span></pre></div>             </td>           </tr>                               <tr id="section-2">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-2">&#182;</a>               </div>               <p>--></p>

<script src='../js/testharness.js'></script>

<script src='../js/move-log.js'></script>

<script src='../using-testharness.js'></script>

<p><link rel='stylesheet' href='../js/testharness.css'>
This document provides a tutorial for W3C's test framework, known as <code>testharness.js</code>,
which you can download from <a href="https://github.com/jgraham/testharness.js">its GitHub repository</a>.</p>

<p>This tutorial does not assume that you are necessarily familiar with other test frameworks,
but it does expect you to be reasonably proficient with JavaScript (since JavaScript APIs is
what one tests using this).</p>

<p>If you are familiar with other test frameworks such as <a href="http://docs.jquery.com/QUnit">QUnit</a>,
<a href="http://visionmedia.github.com/mocha/">Mocha</a>, or <a href="http://pivotal.github.com/jasmine/">Jasmine</a>
then you should find your way around here relatively easily.</p>

<p>Indeed, <code>testharness.js</code> is not very different from those, though it does have a number of
features that make it particularly well suited to testing APIs implemented by the browser
and exposed <em>to</em> JavaScript rather than those created in JavaScript directly.</p>

<p>This tutorial is actually designed to be runnable. The code you see in the right column
is described on the left, but it is also ran as soon as you load this page. If you scroll
to the bottom, you will see the results of this run.</p>

<p>This documentation was generated using <a href="http://jashkenas.github.com/docco/">Docco</a>, and the
idea behind the way in which it is done is shamelessly stolen from the
<a href="http://pivotal.github.com/jasmine/">Jasmine Documentation</a>.</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-3">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-3">&#182;</a>               </div>               <h2>Getting Started</h2>

<p>Let's get started with this code! The first thing you need to do to load <code>testharness.js</code>
is to include it from a <code>script</code> element in the usual way. You can either download your
own copy and set it up locally whichever way you want, or if you're writing a test for
a W3C service you should just point to the W3C copy:</p>

<pre><code>  &lt;script src='http://w3c-test.org/resources/testharness.js'&gt;&lt;/script&gt;
  &lt;script src='http://w3c-test.org/resources/testharnessreport.js'&gt;&lt;/script&gt;
</code></pre>

<p>At which point you might rightfully ask why there are two files. The reason for this is
simple: the first one is the actual implementations, and the second one is empty. Why
include an empty file? The idea is that when a specific vendor checks your test suite
out, they can override the content of the <code>testharnessreport.js</code> file. This makes it
possible for them to integrate running the test suite with whatever test reporting tool
they may be using internally.</p>

<p>If you wish the HTML page into which your tests are being run to render results in a nice
and convenient table, you should include an HTML element with ID <code>log</code> where you would
like those results to appear. See the bottom of this document for an example.</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-4">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-4">&#182;</a>               </div>               <h2>Basic Testing</h2>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-5">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-5">&#182;</a>               </div>               <p>The most basic usage relies on the <code>test()</code> function, which takes a function with the code
to run, a name of the test, and optionally a literal object providing some additional
options.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_true</span><span class="p">(</span><span class="kc">true</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;True really is true&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-6">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-6">&#182;</a>               </div>               <p>The given function <em>must</em> include at least one assertion; conversely, assertions
can only appear in the context of a test function.
A single test may contain multiple assertions, in which case it is considered to be atomic.
That is to say that a single failed assertion will fail the test, whereas all are required
to pass for the whole test to pass. A document can contain as many tests as you wish it
to.</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-7">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-7">&#182;</a>               </div>               <p>The example here contains two assertions, both of which pass.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_true</span><span class="p">(</span><span class="kc">true</span><span class="p">);</span>
    <span class="nx">assert_false</span><span class="p">(</span><span class="kc">false</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Truth is what you believe it to be&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-8">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-8">&#182;</a>               </div>               <p>But in this example one passes, while the other fails. This causes the entire test to be reported
as a failure.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_true</span><span class="p">(</span><span class="kc">true</span><span class="p">);</span>
    <span class="nx">assert_false</span><span class="p">(</span><span class="kc">true</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;All opinions are equally valid.&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-9">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-9">&#182;</a>               </div>               <p>In addition to a function and a name, <code>test()</code> can also accept a third parameter being a dictionary
of options. Most of those options are documented in the <a href="#metadata">Including Metadata</a> section
below.</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-10">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-10">&#182;</a>               </div>               <p>The only general-purpose option that you can use is <code>timeout</code>. It takes a number of milliseconds
(defaulting to 1000, which the brightest at maths amongst you will have recognised as being equivalent
to one second). If the content of the test takes longer than <code>timeout</code> to run then the test is aborted
and counted as a fail. Since some processing can take longer than one second to run, especially if
you are performing a complex test in a low-end environment (e.g. a basic mobile phone) it can at
times be useful to increase this limit as exemplified here.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="cm">/* do something long and slow here */</span>
    <span class="nx">assert_true</span><span class="p">(</span><span class="kc">true</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Long operation is successful&quot;</span><span class="p">,</span> <span class="p">{</span> <span class="nx">timeout</span><span class="o">:</span> <span class="mi">5000</span> <span class="p">});</span></pre></div>             </td>           </tr>                               <tr id="section-11">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-11">&#182;</a>               </div>               <p>Note that you do not want to use this for asynchronous operations (if only because it won't work).
For those, see the section dedicated to that topic below.</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-12">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-12">&#182;</a>               </div>               <h2>Included Assertions</h2>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-13">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-13">&#182;</a>               </div>               <p>There is a good choice of assertions available by default. Most of those take the form
<code>assert_something(actual, expected, description)</code> but several have different signatures.
When a <code>description</code> is part of the signature, it is optional and a string intended for
human consumption in order to provide a better description of which assertion exactly
failed. If you don't provide it, you will get a default error message instead.</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-14">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-14">&#182;</a>               </div>               <p><code>assert_true(actual, description)</code> checks that <code>actual</code> is <em>strictly</em> equal to <code>true</code>,
which is to say that it <em>has</em> to be the JavaScript <code>true</code> value and not just someting
that evaluates as "truthy" such as <code>1</code> or <code>"dahut"</code>.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_true</span><span class="p">(</span><span class="kc">true</span><span class="p">,</span> <span class="s2">&quot;Truth is true&quot;</span><span class="p">);</span>
    <span class="nx">assert_true</span><span class="p">(</span><span class="mi">1</span> <span class="o">===</span> <span class="mi">1</span><span class="p">,</span> <span class="s2">&quot;One is really one&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Simple checks on truth&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-15">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-15">&#182;</a>               </div>               <p><code>assert_false(actual, description)</code> is the same as <code>assert_true</code> but in reverse. It
has the same strictness about its <code>actual</code> being JavaScript's <code>false</code> and not just
"falsy" (e.g. 0, null)</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_false</span><span class="p">(</span><span class="kc">false</span><span class="p">,</span> <span class="s2">&quot;Falsity is false&quot;</span><span class="p">);</span>
    <span class="nx">assert_false</span><span class="p">(</span><span class="mi">1</span> <span class="o">===</span> <span class="mi">0</span><span class="p">,</span> <span class="s2">&quot;One is not zero&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Simple checks on falsity&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-16">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-16">&#182;</a>               </div>               <p><code>assert_equals(actual, expected, description)</code> checks that <code>actual</code> and <code>expected</code>
have the same value (without necessarily being the same object). Note that this
comparison is strict and that you should not rely on
whatever automatic type conversions that JavaScript may perform on comparisons.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_equals</span><span class="p">(</span><span class="s2">&quot;dahut&quot;</span><span class="p">,</span> <span class="s2">&quot;da&quot;</span> <span class="o">+</span> <span class="s2">&quot;hut&quot;</span><span class="p">,</span> <span class="s2">&quot;String concatenation&quot;</span><span class="p">);</span>
    <span class="nx">assert_equals</span><span class="p">(</span><span class="mi">42</span><span class="p">,</span> <span class="mi">6</span> <span class="o">*</span> <span class="mi">7</span><span class="p">,</span> <span class="s2">&quot;The ultimate answer&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Simple checks on equality&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-17">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-17">&#182;</a>               </div>               <p><code>assert_not_equals(actual, expected, description)</code> is the reverse of <code>assert_equals</code>
and checks that its /actual/ and /expected/ are not the same. The same caveat on
comparison strictness applies, so that values that may seem very similar are still
not equal.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_not_equals</span><span class="p">(</span><span class="s2">&quot;dahut&quot;</span><span class="p">,</span> <span class="s2">&quot;myth&quot;</span><span class="p">,</span> <span class="s2">&quot;String comparison&quot;</span><span class="p">);</span>
    <span class="nx">assert_not_equals</span><span class="p">(</span><span class="mi">42</span><span class="p">,</span> <span class="s2">&quot;42&quot;</span><span class="p">,</span> <span class="s2">&quot;The ultimate answer&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Simple checks on unequality&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-18">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-18">&#182;</a>               </div>               <p><code>assert_in_array(actual, expected, description)</code> checks that <code>actual</code> is in the array
provided in <code>expected</code>. Any odd member will do, but note that it will not recurse
into the array if it is multidimensional.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_in_array</span><span class="p">(</span><span class="s2">&quot;dahut&quot;</span><span class="p">,</span>
                    <span class="s2">&quot;chupacabra dahut unicorn&quot;</span><span class="p">.</span><span class="nx">split</span><span class="p">(</span><span class="s2">&quot; &quot;</span><span class="p">),</span>
                    <span class="s2">&quot;Dahut hunting&quot;</span><span class="p">);</span>
    <span class="nx">assert_in_array</span><span class="p">(</span><span class="mi">2017</span><span class="p">,</span> <span class="p">[</span><span class="mi">42</span><span class="p">,</span> <span class="mi">47</span><span class="p">,</span> <span class="mi">62</span><span class="p">,</span> <span class="mi">2017</span><span class="p">]</span> <span class="p">,</span> <span class="s2">&quot;Lottery&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Simple checks on membership&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-19">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-19">&#182;</a>               </div>               <p><code>assert_array_equals(actual, expected, description)</code> takes an array for both <code>actual</code>
and <code>expected</code>, and validates that they have the same length and that each item is
<code>assert_equals</code> its corresponding member in the other array. Just like the previous
assertion, this is unidimensional.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_array_equals</span><span class="p">([</span><span class="s2">&quot;chupacabra&quot;</span><span class="p">,</span> <span class="s2">&quot;dahut&quot;</span><span class="p">,</span> <span class="s2">&quot;unicorn&quot;</span><span class="p">],</span>
                        <span class="s2">&quot;chupacabra dahut unicorn&quot;</span><span class="p">.</span><span class="nx">split</span><span class="p">(</span><span class="s2">&quot; &quot;</span><span class="p">),</span>
                        <span class="s2">&quot;Dahut hunting&quot;</span><span class="p">);</span>
    <span class="nx">assert_array_equals</span><span class="p">([</span><span class="mi">4</span><span class="p">,</span> <span class="mi">9</span><span class="p">,</span> <span class="mi">16</span><span class="p">],</span>
                        <span class="p">[</span><span class="mi">2</span><span class="p">,</span> <span class="mi">3</span><span class="p">,</span> <span class="mi">4</span><span class="p">].</span><span class="nx">map</span><span class="p">(</span><span class="kd">function</span> <span class="p">(</span><span class="nx">x</span><span class="p">)</span> <span class="p">{</span> <span class="k">return</span> <span class="nx">x</span> <span class="o">*</span> <span class="nx">x</span><span class="p">;</span> <span class="p">}),</span>
                        <span class="s2">&quot;Square&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks on identical membership&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-20">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-20">&#182;</a>               </div>               <p><code>assert_approx_equals(actual, expected, epsilon, description)</code> takes a numerical <code>actual</code> value
and checks that it is within <code>epsilon</code> of <code>expected</code>. This is notably useful for floating point
calculations in which you know that some drift may occur and you need to check that the outcome
is within a given ballpark &mdash; but it can also be used in other cases.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_approx_equals</span><span class="p">(</span><span class="nb">Math</span><span class="p">.</span><span class="nx">PI</span><span class="p">,</span> <span class="mf">3.14</span><span class="p">,</span> <span class="mf">0.01</span><span class="p">,</span> <span class="s2">&quot;Roughly circular&quot;</span><span class="p">);</span>
    <span class="nx">assert_approx_equals</span><span class="p">(</span><span class="mi">42</span><span class="p">,</span> <span class="mi">47</span><span class="p">,</span> <span class="mi">5</span><span class="p">,</span> <span class="s2">&quot;47 is almost 42&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks on epsilon equality&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-21">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-21">&#182;</a>               </div>               <p><code>assert_regexp_match(actual, expected, description)</code> checks that <code>actual</code> matches the <code>expected</code>
regular expression. The latter can be as simple or complex as you wish to make it, and can be
created with flags.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_regexp_match</span><span class="p">(</span><span class="nb">document</span><span class="p">.</span><span class="nx">title</span><span class="p">,</span>
                        <span class="sr">/^\w{5}-\w{10,12}\.js$/</span><span class="p">,</span>
                        <span class="s2">&quot;That&#39;s my title&quot;</span><span class="p">);</span>
    <span class="nx">assert_regexp_match</span><span class="p">(</span><span class="s2">&quot;A&quot;</span><span class="p">,</span> <span class="sr">/a/i</span><span class="p">,</span> <span class="s2">&quot;Matching lowercase&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks using regular expressions&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-22">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-22">&#182;</a>               </div>               <p><code>assert_own_property(object, property_name, description)</code> checks that <code>object</code> has a property that
is truly its own (as opposed to inherited down the prototype chain). JavaScripters will recognise
this as checking <code>hasOwnProperty</code>. If you don't know about this important method, you can
<a href="https://developer.mozilla.org/en/JavaScript/Reference/Global_Objects/Object/hasOwnProperty">read up about it on MDN</a>.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="kd">var</span> <span class="nx">gollum</span> <span class="o">=</span> <span class="p">{</span> <span class="nx">ring</span><span class="o">:</span> <span class="s2">&quot;MIIIIINE!!!!&quot;</span> <span class="p">};</span>
    <span class="nx">assert_own_property</span><span class="p">(</span><span class="nx">gollum</span><span class="p">,</span> <span class="s2">&quot;ring&quot;</span><span class="p">,</span> <span class="s2">&quot;Tricksy hobbitses!&quot;</span><span class="p">);</span>
    <span class="cm">/* this will fail even though `gollum` has `toString`. */</span>
    <span class="nx">assert_own_property</span><span class="p">(</span><span class="nx">gollum</span><span class="p">,</span>
                        <span class="s2">&quot;toString&quot;</span><span class="p">,</span>
                        <span class="s2">&quot;I have that property, but it&#39;ssss not mine.&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks for property ownership&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-23">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-23">&#182;</a>               </div>               <p><code>assert_inherits(object, property_name, description)</code> complements <code>assert_own_property</code> in that it
similarly checks that the attribute is available on the object, but asserts that it is <em>not</em> the
object's own property but rather has been inherited down the prototype chain.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="kd">var</span> <span class="nx">gollum</span> <span class="o">=</span> <span class="p">{</span> <span class="nx">ring</span><span class="o">:</span> <span class="s2">&quot;MIIIIINE!!!!&quot;</span> <span class="p">};</span>
    <span class="cm">/* this will succeed here */</span>
    <span class="nx">assert_inherits</span><span class="p">(</span><span class="nx">gollum</span><span class="p">,</span>
                    <span class="s2">&quot;toString&quot;</span><span class="p">,</span>
                    <span class="s2">&quot;I have that property, but it&#39;ssss not mine.&quot;</span><span class="p">);</span>
    <span class="nx">assert_inherits</span><span class="p">(</span><span class="nx">gollum</span><span class="p">,</span>
                    <span class="s2">&quot;hasOwnProperty&quot;</span><span class="p">,</span>
                    <span class="s2">&quot;This one works too.&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks for property inheritance&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-24">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-24">&#182;</a>               </div>               <p><code>assert_idl_attribute(object, attribute_name, description)</code> is the same as
<code>assert_inherits</code> and simply aliases it. For clarity, you may be better off sticking
to the previous one.</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-25">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-25">&#182;</a>               </div>               <p><code>assert_readonly(object, property_name, description)</code> checks that the given <code>property_name</code>
on <code>object</code> is properly read-only and therefore cannot be set.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_readonly</span><span class="p">(</span><span class="nb">document</span><span class="p">,</span> <span class="s2">&quot;nodeType&quot;</span><span class="p">,</span> <span class="s2">&quot;You cannot change nodeType.&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks for attribute readonlyness&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-26">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-26">&#182;</a>               </div>               <p><code>assert_throws(code, func, description)</code> is a powerful way of checking that code throws when
and how you expect it to, knowing that the code in <code>func</code> is what must trigger the exception.
This assertion works differently depending on what you pass for <code>code</code>.</p>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                               <tr id="section-27">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-27">&#182;</a>               </div>               <p>If <code>code</code> is <code>null</code>, then any old exception will do (this is not a particularly recommended check as
the others are more useful).</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_throws</span><span class="p">(</span><span class="s2">&quot;HierarchyRequestError&quot;</span><span class="p">,</span>
                  <span class="kd">function</span> <span class="p">()</span> <span class="p">{</span> <span class="nb">document</span><span class="p">.</span><span class="nx">appendChild</span><span class="p">(</span><span class="nb">document</span><span class="p">);</span> <span class="p">},</span>
                  <span class="s2">&quot;Specific DOM exception.&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks for exceptions (null)&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-28">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-28">&#182;</a>               </div>               <p>If <code>code</code> is any kind of object, then its <code>name</code> attribute is checked. That attribute must match the
<code>name</code> attribute on the exception being thrown. This means that you can pass a specific <code>DOMError</code>
object here and have it match if it's what is being thrown.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_throws</span><span class="p">({</span> <span class="nx">name</span><span class="o">:</span> <span class="s2">&quot;Bad Kitten!&quot;</span> <span class="p">},</span>
                  <span class="kd">function</span> <span class="p">()</span> <span class="p">{</span> <span class="k">throw</span> <span class="p">{</span> <span class="nx">name</span><span class="o">:</span> <span class="s2">&quot;Bad Kitten!&quot;</span><span class="p">};</span> <span class="p">},</span>
                  <span class="s2">&quot;Any exception with the right name.&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks for exceptions (object)&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-29">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-29">&#182;</a>               </div>               <p>If <code>code</code> is a string then it must be one of the commonly recognised <code>DOMError</code> names, and it checks
that <code>func</code> throws the corresponding <code>DOMError</code>. For compatibility with older browsers, the old
exception contacts are supported and mapped to the newer name; so for instance you can use
<code>WRONG_DOCUMENT_ERR</code> to mean <code>WrongDocumentError</code>. The latter style is preferred however.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_throws</span><span class="p">(</span><span class="s2">&quot;HierarchyRequestError&quot;</span><span class="p">,</span>
                  <span class="kd">function</span> <span class="p">()</span> <span class="p">{</span> <span class="nb">document</span><span class="p">.</span><span class="nx">appendChild</span><span class="p">(</span><span class="nb">document</span><span class="p">);</span> <span class="p">},</span>
                  <span class="s2">&quot;Specific DOM exception.&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks for exceptions (string)&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-30">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-30">&#182;</a>               </div>               <p><code>assert_object_equals(actual, expected, description)</code> checks that two objects are equal by deep-walking
them side by side and making sure that they have the same fields and that those fields have the same
values. It is still considered somewhat experimental.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_object_equals</span><span class="p">({</span> <span class="nx">foo</span><span class="o">:</span> <span class="s2">&quot;bar&quot;</span> <span class="p">},</span>
                         <span class="p">{</span> <span class="nx">foo</span><span class="o">:</span> <span class="s2">&quot;bar&quot;</span> <span class="p">},</span>
                         <span class="s2">&quot;Simple objects.&quot;</span><span class="p">);</span>
    <span class="nx">assert_object_equals</span><span class="p">({</span> <span class="nx">top</span><span class="o">:</span> <span class="s2">&quot;here&quot;</span><span class="p">,</span> <span class="nx">kids</span><span class="o">:</span> <span class="p">{</span> <span class="nx">list</span><span class="o">:</span> <span class="p">[</span><span class="s2">&quot;stuff&quot;</span><span class="p">,</span> <span class="p">{</span> <span class="nx">leaf</span><span class="o">:</span> <span class="kc">true</span> <span class="p">}</span> <span class="p">]}</span> <span class="p">},</span>
                         <span class="p">{</span> <span class="nx">top</span><span class="o">:</span> <span class="s2">&quot;here&quot;</span><span class="p">,</span> <span class="nx">kids</span><span class="o">:</span> <span class="p">{</span> <span class="nx">list</span><span class="o">:</span> <span class="p">[</span><span class="s2">&quot;stuff&quot;</span><span class="p">,</span> <span class="p">{</span> <span class="nx">leaf</span><span class="o">:</span> <span class="kc">true</span> <span class="p">}</span> <span class="p">]}</span> <span class="p">},</span>
                         <span class="s2">&quot;Simple objects.&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Checks object deep-equality&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-31">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-31">&#182;</a>               </div>               <p><code>assert_unreached(description)</code> is a very simple assertion the role of which is to
check that some code is indeed unreachable. It only takes a description, and simply
always throws its hands up in disgust whenever it is called. Opposite here you
can see a case in which it is successful (since untouched).</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="k">if</span> <span class="p">(</span><span class="kc">true</span><span class="p">)</span> <span class="k">return</span> <span class="s2">&quot;where you came from&quot;</span><span class="p">;</span>
    <span class="nx">assert_unreached</span><span class="p">(</span><span class="s2">&quot;Can&#39;t Touch This&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Simple check on unreachability&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-32">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-32">&#182;</a>               </div>               <p>Whereas this one fails because the code reaches it.</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="nx">test</span><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    <span class="nx">assert_unreached</span><span class="p">(</span><span class="s2">&quot;Reaching where no coder has reached before&quot;</span><span class="p">);</span>
<span class="p">},</span> <span class="s2">&quot;Failed check on unreachability&quot;</span><span class="p">);</span></pre></div>             </td>           </tr>                               <tr id="section-33">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-33">&#182;</a>               </div>               <h2>Asynchronous Testing</h2>             </td>             <td class="code">               <div class="highlight"><pre><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    
<span class="p">}());</span></pre></div>             </td>           </tr>                               <tr id="section-34">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-34">&#182;</a>               </div>               <p><a name='metadata'></a></p>

<h2>Including Metadata</h2>             </td>             <td class="code">               <div class="highlight"><pre><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    
<span class="p">}());</span></pre></div>             </td>           </tr>                               <tr id="section-35">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-35">&#182;</a>               </div>               <h2>Advanced Usage</h2>

<p>format_value</p>             </td>             <td class="code">               <div class="highlight"><pre><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    
<span class="p">}());</span></pre></div>             </td>           </tr>                               <tr id="section-36">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-36">&#182;</a>               </div>               <h2>Generating Tests</h2>             </td>             <td class="code">               <div class="highlight"><pre><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    
<span class="p">}());</span></pre></div>             </td>           </tr>                               <tr id="section-37">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-37">&#182;</a>               </div>               <h2>Writing Your Own Assertions</h2>             </td>             <td class="code">               <div class="highlight"><pre><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    
<span class="p">}());</span></pre></div>             </td>           </tr>                               <tr id="section-38">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-38">&#182;</a>               </div>               <h2>Callbacks</h2>             </td>             <td class="code">               <div class="highlight"><pre><span class="p">(</span><span class="kd">function</span> <span class="p">()</span> <span class="p">{</span>
    
<span class="p">}());</span></pre></div>             </td>           </tr>                               <tr id="section-39">             <td class="docs">               <div class="pilwrap">                 <a class="pilcrow" href="#section-39">&#182;</a>               </div>               <h2>Results</h2>

<p>As promised this is a self-runnable document that includes the results for the test suite
specified by the code in the right column above. You can see the results below:</p>

<div id='log'></div>             </td>             <td class="code">               <div class="highlight"><pre></pre></div>             </td>           </tr>                </tbody>     </table>   </div> </body> </html> 
